requirements.txt
pyTelegramBotAPI==4.14.0
requests==2.31.0
python-dotenv==1.0.0
openai>=1.6.0
colorama==0.4.6
Flask==2.3.3
pandas==2.1.1
openpyxl==3.1.2
xlrd==2.0.1
numpy==1.24.3
Werkzeug==2.3.7
pdfplumber==0.10.3 